Extract relay hint ownership
