+++ web/src/lib/RelayList.ts
@@ -2,7 +2,8 @@ import { createRxBackwardReq, latestEach, uniq } from 'rx-nostr';
 import { kinds as Kind } from 'nostr-tools';
 import type * as Nostr from 'nostr-typedef';
 import type { pubkey } from './Types';
-import { rxNostr, tie } from './timelines/MainTimeline';
+import { rxNostr } from '$lib/nostr/client';
+import { tie } from '$lib/nostr/relay-hints';
 import { cacheFolloweeReplaceableEvent, followeeEventCache } from './cache/Events';
 
 export class RelayList {
